# source/swagger.yaml
openapi: 3.0.2
info:
  title: Post articles API
  description: The API allows you to register, search, change, delete user accounts. After authentication, users can create, search, edit, delete articles.
  version: 1.0.0
  contact:
    name: Molochkov Sergey
    email: wwwsergwww@mail.ru
  license:
    name: ISC
servers:
  - url: /
...
paths:
...
  /users:
    post:
      tags:
        - User
      description: Creates user
      requestBody:
        description: User account that needs to be added to the service
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSchema'
      responses:
        '200':
          description: Create new user account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserView'
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationError'
                  - $ref: '#/components/schemas/DuplicateValueError'
                  - $ref: '#/components/schemas/CreationError'
        '429':
          $ref: '#/components/responses/TooManyPequestsError'
    get:
      tags:
        - User
      description: Get all users
      responses:
        '200':
          description: Gets an array of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserShortView'
        '403':
          $ref: '#/components/responses/AuthorizeError'
      security:
        - JWT: []
...
  /articles:
    post:
      tags:
        - Article
      description: Creates article
      requestBody:
        description: Article that needs to be added to the service
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleSchema'
      responses:
        '200':
          description: Create new article
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArticleView'
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationError'
                  - $ref: '#/components/schemas/DuplicateValueError'
                  - $ref: '#/components/schemas/CreationError'
        '403':
          $ref: '#/components/responses/AuthorizeError'
      security:
        - JWT: []
    get:
      tags:
        - Article
      description: Get all articles
      responses:
        '200':
          description: Gets an array of articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArticleShortView'
      security:
        - JWT: []
...
components:
  schemas:
...
    UserSchema:
      required:
        - name
        - sex
        - birthDay
        - phone
        - email
        - password
      properties:
        name:
          type: string
          description: Value, should have a minimum 8 characters, maximum 25 characters,  contain a-zA-Z0-9
          example: Tom
        sex:
          type: string
          enum:
            - male
            - female
          example: male
        birthDay:
          type: string
          format: dateTime
          description: Registration is available to users over the age of 18. Valid values are from 1900-01-01.
          example: 1988-01-01
        phone:
          type: string
          description: Value should be like at +7(8)-nnn-nnn-nn-nn
          example: +7-123-456-78-90
        email:
          type: string
          description: Value should be a valid email
          example: mail@mail.ru
        password:
          type: string
          description: Value should have a minimum 8 characters, at least one letter, one number and one special character
          example: Sw331@111
    ArticleSchema:
      required:
        - theme
      properties:
        theme:
          type: string
          description: Value, should have a minimum 3 characters, maximum 100 characters
          example: Some theme
        content:
          type: string
          example: Very interesting text
    UserView:
      properties:
        id:
          type: string
          example: 62de778c48bdbcddfa21ff61
        name:
          type: string
          example: Tom
        sex:
          type: string
          enum:
            - male
            - female
          example: male
        age:
          type: number
          example: 34
        phone:
          type: string
          example: +7-123-456-78-90
        email:
          type: string
          example: mail@mail.ru
        articles:
          type: array
          items:
            $ref: '#/components/schemas/ArticleShortView'
    UserShortView:
      properties:
        id:
          type: string
          example: 62de778c48bdbcddfa21ff61
        name:
          type: string
          example: Tom
        sex:
          type: string
          enum:
            - male
            - female
          example: male
        age:
          type: number
          example: 34
        phone:
          type: string
          example: +7-123-456-78-90
        email:
          type: string
          example: mail@mail.ru
    ArticleView:
      properties:
        id:
          type: string
          example: 62de778c48bdbcddfa21ff61
        author:
          properties:
            id:
              type: string
              example: 62de778c48bdbcddfa21ff61
            name:
              type: string
              example: Bob
        theme:
          type: string
          example: Some theme
        content:
          type: string
          example: Very interesting text
        check:
          type: boolean
          description: Article status
          example: true
        created:
          type: string
          format: date-time
          example: '2022-07-25T10:59:33.099Z'
        modified:
          type: string
          format: date-time
          example: '2022-07-25T10:59:33.099Z'
    ArticleShortView:
      properties:
        id:
          type: string
          example: 62de778c48bdbcddfa21ff61
        author:
          type: string
          example: Bob
        theme:
          type: string
          example: Some theme
        check:
          type: boolean
          description: Article status
          example: true
        created:
          type: string
          format: date-time
          example: '2022-07-25T10:59:33.099Z'
        modified:
          type: string
          format: date-time
          example: '2022-07-25T10:59:33.099Z'
    ValidationError:
      properties:
        message:
          type: string
          example: 'error.message'
...
    DuplicateValueError:
      properties:
        message:
          type: string
          example: 'Duplicate value error in error.keyValue'
    CreationError:
      properties:
        message:
          type: string
          example: 'Incorrect payload'
  responses:
    TooManyPequestsError:
      description: Too Many Pequests
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
                example: 'Too many authentications created from this IP, please try again after an X seconds'
...
    AuthorizeError:
      description: Forbidden
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
                example: 'That may be, but you have no right to access it'
...
  securitySchemes:
    JWT:
      type: apiKey
      in: header
      name: access_token